+++ packages/shared/src/types/index.ts
@@ -1,4 +1,3 @@
-export * as LocalRestAPI from "./plugin-local-rest-api";
 export * as SmartConnections from "./plugin-smart-connections";
 export * as Templater from "./plugin-templater";
 export * from "./prompts";